python-fastapi/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
httpx==0.28.1
google-auth>=2.0.0
